+++ scripts/proxyArgs.ts
@@ -1,0 +1,7 @@
+const args = [
+  "0x800572D7D4106ca0b1de7f8c3185D9A3Aca6b72a", // implementation
+  "0x92371F3c1Bda1fB4F67a7984FDa54d60b0e6c6Ef", // initial owner
+  "0x8129fc1c" // init data
+];
+
+export default args;